Feature: Add localization to application title, rename localization plugin
