+++ src/locales/en-US.json
@@ -4,7 +4,10 @@
 		"home": {
 			"title": "Home"
 		},
-		"test": {
+		"session": {
+			"title": "Session"
+		},
+		"quiz": {
 			"title": "Quiz"
 		}
 	}
