+++ src/main.ts
@@ -4,11 +4,11 @@ import { createApp } from 'vue'
 
 import App from '@/App.vue'
 import router from '@/router'
-import { localizationPlugin } from '@/plugins'
+import { I18nPlugin } from '@/plugins'
 
 const app = createApp(App)
 
 app.use(router)
-app.use(localizationPlugin)
+app.use(I18nPlugin)
 
 app.mount('#app')
