+++ src/plugins/I18nPlugin.ts
