+++ src/views/QuizView.vue
@@ -1,0 +1,3 @@
+<template>
+    <div>{{ $t('app.quiz.title') }}</div>
+</template>
