+++ src/views/SessionView.vue
@@ -1,0 +1,3 @@
+<template>
+    <div>{{ $t('app.session.title') }}</div>
+</template>
