--- src/views/TestView.vue
@@ -1,3 +1,0 @@
-<template>
-    <div>Test</div>
-</template>
